+++ test/basic/error.c
@@ -76,7 +76,7 @@ int main(int argc, char *argv[])
         { "ABT_ERR_INV_ARG", ABT_ERR_INV_ARG },
     };
 
-    for (i = 0; i < sizeof(error_pairs) / sizeof(error_pairs[0]); i++) {
+    for (i = 0; i < (int)(sizeof(error_pairs) / sizeof(error_pairs[0])); i++) {
         char str[256];
         ret = ABT_error_get_str(error_pairs[i].code, str, NULL);
         ATS_ERROR(ret, "ABT_error_get_str");
